Added rays

+++ MARGE/src/includes/ray.cpp
@@ -3,7 +3,7 @@
 ray::ray() : origin(point3(0, 0, 0)), direction(vector3(0, 0, 0)) {};
 ray::ray(const point3& origin, const vector3& direction) : origin(origin), direction(direction) {};
 
-point3 ray::at(float& dist) 
+point3 ray::at(double& dist) 
 {
     return this->origin + (this->direction * dist);
 }
